+++ platform/android/dnsproxy/lib/src/main/cpp/jni_defs.h
@@ -24,3 +24,4 @@
 #define FQN_DNSSTAMP_PROTOTYPE FQN_DNSSTAMP "$ProtoType"
 #define FQN_DNSSTAMP_INFORMAL_PROPERTIES FQN_DNSSTAMP "$InformalProperties"
 #define FQN_DNS_MESSAGE_INFO "com/adguard/dnslibs/proxy/DnsMessageInfo"
+#define FQN_BLOCKING_REASON "com/adguard/dnslibs/proxy/DnsBlockingReason"